+++ backend/canisters/local_user_index/api/src/main.rs
@@ -21,6 +21,7 @@ fn main() {
     generate_ts_method!(local_user_index, bot_create_channel);
     generate_ts_method!(local_user_index, bot_delete_channel);
     generate_ts_method!(local_user_index, bot_delete_messages);
+    generate_ts_method!(local_user_index, bot_invite_users);
     generate_ts_method!(local_user_index, bot_send_message);
     generate_ts_method!(local_user_index, bot_subscribe_to_events);
     generate_ts_method!(local_user_index, install_bot);